split MultiDepositParser into multiple files
smaller files give faster compile time
