+++ src/main/scala/nl/knaw/dans/easy/multideposit/parser/ParserUtils.scala
@@ -1,0 +1,97 @@
+package nl.knaw.dans.easy.multideposit.parser
+
+import java.util.Locale
+
+import nl.knaw.dans.easy.multideposit._
+import nl.knaw.dans.easy.multideposit.model._
+import nl.knaw.dans.lib.error._
+
+import scala.language.implicitConversions
+import scala.util.{ Failure, Success, Try }
+
+trait ParserUtils {
+
+  def getRowNum(row: DatasetRow): Int = row("ROW").toInt
+
+  def extractNEL[T](rows: DatasetRows)(f: (=> Int) => DatasetRow => Option[Try[T]]): Try[NonEmptyList[T]] = {
+    rows.flatMap(row => f(getRowNum(row))(row)).collectResults.map(_.toList)
+  }
+
+  def extractNEL(rows: DatasetRows, rowNum: Int, name: MultiDepositKey): Try[NonEmptyList[String]] = {
+    rows.flatMap(_.find(name)) match {
+      case Seq() => Failure(ParseException(rowNum, s"There should be at least one non-empty value for $name"))
+      case xs => Try { xs.toList }
+    }
+  }
+
+  def extractList[T](rows: DatasetRows)(f: (=> Int) => DatasetRow => Option[Try[T]]): Try[List[T]] = {
+    rows.flatMap(row => f(getRowNum(row))(row)).collectResults.map(_.toList)
+  }
+
+  def extractList(rows: DatasetRows, name: MultiDepositKey): List[String] = {
+    rows.flatMap(_.find(name)).toList
+  }
+
+  def atMostOne[T](rowNum: => Int, columnNames: => NonEmptyList[MultiDepositKey])(values: List[T]): Try[Option[T]] = {
+    values.distinct match {
+      case Nil => Success(None)
+      case t :: Nil => Success(Some(t))
+      case ts if columnNames.size == 1 => Failure(ParseException(rowNum, "Only one row is allowed " +
+        s"to contain a value for the column '${ columnNames.head }'. Found: ${ ts.mkString("[", ", ", "]") }"))
+      case ts => Failure(ParseException(rowNum, "Only one row is allowed to contain a value for " +
+        s"these columns: ${ columnNames.mkString("[", ", ", "]") }. Found: ${ ts.mkString("[", ", ", "]") }"))
+    }
+  }
+
+  def exactlyOne[T](rowNum: => Int, columnNames: => NonEmptyList[MultiDepositKey])(values: List[T]): Try[T] = {
+    values.distinct match {
+      case t :: Nil => Success(t)
+      case Nil if columnNames.size == 1 => Failure(ParseException(rowNum, "One row has to contain " +
+        s"a value for the column: '${ columnNames.head }'"))
+      case Nil => Failure(ParseException(rowNum, "One row has to contain a value for these " +
+        s"columns: ${ columnNames.mkString("[", ", ", "]") }"))
+      case ts if columnNames.size == 1 => Failure(ParseException(rowNum, "Only one row is allowed " +
+        s"to contain a value for the column '${ columnNames.head }'. Found: ${ ts.mkString("[", ", ", "]") }"))
+      case ts => Failure(ParseException(rowNum, "Only one row is allowed to contain a value for " +
+        s"these columns: ${ columnNames.mkString("[", ", ", "]") }. Found: ${ ts.mkString("[", ", ", "]") }"))
+    }
+  }
+
+  def checkValidChars(rowNum: => Int, column: => MultiDepositKey, value: String): Try[String] = {
+    val invalidCharacters = "[^a-zA-Z0-9_-]".r.findAllIn(value).toSeq.distinct
+    if (invalidCharacters.isEmpty) Success(value)
+    else Failure(ParseException(rowNum, s"The column '$column' contains the following invalid characters: ${ invalidCharacters.mkString("{", ", ", "}") }"))
+  }
+
+  def missingRequired[T](rowNum: Int, row: DatasetRow, required: Set[String]): Failure[T] = {
+    val blankRequired = row.collect { case (key, value) if value.isBlank && required.contains(key) => key }
+    val missingColumns = required.diff(row.keySet)
+    val missing = blankRequired.toSet ++ missingColumns
+    require(missing.nonEmpty, "the list of missing elements is supposed to be non-empty")
+    Failure(ParseException(rowNum, s"Missing value(s) for: ${ missing.mkString("[", ", ", "]") }"))
+  }
+
+  def isValidISO639_1Language(lang: String): Boolean = {
+    val b0: Boolean = lang.length == 2
+    val b1: Boolean = new Locale(lang).getDisplayLanguage.toLowerCase != lang.toLowerCase
+
+    b0 && b1
+  }
+
+  private lazy val iso639_2Languages = Locale.getISOLanguages.map(new Locale(_).getISO3Language).toSet
+
+  def iso639_2Language(columnName: MultiDepositKey)(rowNum: => Int)(row: DatasetRow): Option[Try[String]] = {
+    row.find(columnName)
+      .map(lang => {
+        // Most ISO 639-2/T languages are contained in the iso639_2Languages Set.
+        // However, some of them are not and need to be checked using the second predicate.
+        // The latter also allows to check ISO 639-2/B language codes.
+        lazy val b0 = lang.length == 3
+        lazy val b1 = iso639_2Languages.contains(lang)
+        lazy val b2 = new Locale(lang).getDisplayLanguage.toLowerCase != lang.toLowerCase
+
+        if (b0 && (b1 || b2)) Success(lang)
+        else Failure(ParseException(rowNum, s"Value '$lang' is not a valid value for $columnName"))
+      })
+  }
+}
